test(core): split timed boss bar coverage
